service/requirements.txt
docker
pika
pika-stubs
python-dotenv
websockets